refactore: props type 분리
