+++ src/app/_types/user.ts
@@ -1,0 +1,11 @@
+// export interface UserInfoType {
+//   userId: string;
+//   userName: string;
+//   userImg: string;
+// }
+
+export interface UserProps {
+  userId: string;
+  userName: string;
+  userImg: string;
+}
